backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# OpenAI
openai==1.3.7

# ASR (Faster-Whisper)
faster-whisper==0.10.0
torch==2.1.1
torchaudio==2.1.1

# Audio processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Resume parsing
pdfminer.six==20221105
python-docx==1.1.0
spacy==3.7.2
unstructured==0.11.6

# Vector database
qdrant-client==1.7.0
sentence-transformers==2.2.2

# AWS
boto3==1.34.0
botocore==1.34.0

# Environment and config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# WebSocket
websockets==12.0

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

